작성자 개념 추가/작성시,수정,삭제 권한 추가

+++ src/main/webapp/jsp/article/detail.jsp
@@ -4,6 +4,7 @@
 
 <%
 Map<String, Object> articleRow = (Map<String, Object>) request.getAttribute("articleRow");
+boolean isLogined = (boolean) request.getAttribute("isLogined");
 %>
 <!DOCTYPE html>
 <html>
@@ -26,8 +27,17 @@ Map<String, Object> articleRow = (Map<String, Object>) request.getAttribute("art
 		내용 :
 		<%=articleRow.get("body")%></div>
 	<div>
+		작성자:
+		<%=articleRow.get("name")%></div>
+	<div>
+	<%
+	if (isLogined) {
+	%>
 		<a href="modify?id=<%=articleRow.get("id")%>">수정</a> <a
 			href="doDelete?id=<%=articleRow.get("id")%>">del</a>
+			<%
+	}
+			%>
 	</div>
 	<div>
 		<a style="color: green" href="list">리스트로 돌아가기</a>
