+++ src/main/webapp/jsp/article/list.jsp
@@ -9,6 +9,7 @@ int cPage = (int) request.getAttribute("page");
 int totalPage = (int) request.getAttribute("totalPage");
 int totalCnt = (int) request.getAttribute("totalCnt");
 int itemsInAPage = (int) request.getAttribute("itemsInAPage");
+boolean isLogined = (boolean) request.getAttribute("isLogined");
 %>
 <!DOCTYPE html>
 <html>
@@ -64,7 +65,13 @@ body {
 		<a href="../home/main">메인으로 이동</a>
 	</div>
 	<div>
+	<%
+				if (isLogined) {
+				%>	
 		<a href="write">글쓰기</a>
+			<%
+				}
+				%>
 	</div>
 
 	<h2>게시물 목록</h2>
@@ -76,8 +83,15 @@ body {
 				<th>번호</th>
 				<th>작성날짜</th>
 				<th>제목</th>
+				<th>작성자</th>
+				<%
+				if (isLogined) {
+				%>
 				<th>수정</th>
 				<th>삭제</th>
+				<%
+				}
+				%>
 			</tr>
 		</thead>
 		<tbody>
@@ -88,8 +102,15 @@ body {
 				<td><%=articleRow.get("id")%></td>
 				<td><%=articleRow.get("regDate")%></td>
 				<td><a href="detail?id=<%=articleRow.get("id")%>"><%=articleRow.get("title")%></a></td>
+				<td><%=articleRow.get("name")%></td>
+				<%
+				if (isLogined) {
+				%>	
 				<td><a href="modify?id=<%=articleRow.get("id")%>">수정</a></td>
 				<td><a href="doDelete?id=<%=articleRow.get("id")%>">del</a></td>
+					<%
+				}
+				%>
 			</tr>
 			<%
 			}
